requirements.txt
coverage==4.0
linecache2==1.0.0
nose==1.3.7
PyYAML==3.12
six==1.10.0
traceback2==1.4.0
unittest2==1.1.0
requests-futures==0.9.7
requests==2.11.1
futures==3.0.5
wheel==0.24.0
parsel==1.0.3
regex==2016.10.22
